changes in structure
