--- .flake8
@@ -1,7 +1,0 @@
-[flake8]
-max-line-length = 99
-ignore = W504, W503, E266
-exclude =
-    .git,
-    __pycache__,
-    __init__.py,
